Book: In Bootstrapping, added --copy option, minor corrections
With this option, ./boostrap will make copies of some of the small
text files (like INSTALL, or some of the files in gnulib/build-aux)
instead of making symbolic links. The symbolic links can be a cause of
confusion in the file lists (with ls) or in backup systems. The issue
is explained in the manual and so the users can ignore this option if
they like.