Development: dir-locals distributed in tar, edited explanations
The .dir-locals.el file was included in the EXTRA_DIST variable of
Makefile.am. It will thus be shipped along with the Gnuastro release
tar-ball in case people want to play with the code before actually
getting the version controlled source.

Some minor edits were also applied to the .dir-locals.el comments. The
extra book explanation about Emacs was put in a cartouche to separate
it from the body of the text and attract only interested readers. Its
text was also edited alittle and some minor additions were made.