All: Number of CPU threads found at runtime
Finished task #13986: The number of threads until now was determined
at runtime instead of through ./configure. As discussed on the
original post to this task, this can cause lots of inefficiencies.

However, the _SC_NPROCESSORS_ONLN or get_nprocs() were not used. In
the Gnulib manual it is said that they are not portable. Gnulib has
suggested using Gnulib's own `num_processors' function (in nproc.h)
for portability which operates really nicely. In the utilities I just
had to change the config.h macro name to a call of this function.

Therefore the --with-numthreads option to configure is no longer
necessary. If someone wants to configure all the programs to use a
fixed number of threads, they can easily specify it in the system-wide
configuration file of each program (before installation for example).