Development: Renamed imported/ to bootstrapped/
The "imported/" name for the outputs of the bootstrapping process was
ambiguous: What is imported? From where? Is it hand written and part of the
utility or library source codes (like "include/")? What is the difference
with "include/" (for a non-programmer)?

So the name of the directory was changed to "bootstrapped/" to be more
clear and directly tell anyone who is looking at the contents of the top
directory that it has something to do with "bootstrap" and
"bootstrap.conf".