Installation: Only local configuration files in `make check'
Until now, when `make check' was run, the utilities would also go
looking into the user and system-wide directories for options. The
`onlydirconf' option is now added to the configuration file of all the
utilities at testing time so the utilities only look into
tests/.gnuastro/ for configuration files. The system was also set in
place to add other such options specific to `make check' in the future
if need be.

Checking user and system configuration files caused problems when the
newer version of a utility has changed old options (renamed, or
removed them). Although this should not happen too often, but in major
releases and during the first few years of Gnuastro (while it is
stabilizing) it might happen. So the checks would fail with the
respective utility complaining of an unrecognized option.