Development: ./bootstrap will not write to .gitignore
When ./bootstrap doesn't find the exact value of its build_aux
variable (we have defined it in bootstrap.conf), it will add it to the
top of the .gitignore file (which we don't want). So the variable
values in configure.ac, bootstrap.conf, and .gitignore are now unified
to avoid this behavior of ./bootstrap, while satisfying our .gitignore
convention for directories.