Development: build_aux in bootstrap.conf corrected
The specified value of build_aux in bootstrap.conf ended in a slash
which would cause `make distcheck' to not find some of its targets. So
to avoid the problem, the slash was removed from the end of this
variable and the other similar variables. Note that as discussed in
two previous commits (57b42c4), we also needed to harmonize this with
.gitignore and configure.ac.