dev: coding conventions (line-width, and Gnulib inclusion)
The suggested 70 character linewidth was arguably too short (the GNU Coding
Standards suggest a line width of 79 or less). But since all the work done
until now was with a 70 character line width, suddenly jumping to 79 would
make the documentation and code sources very ugly! So we opted for a
compromise of 75 character line widths. The corresponding parts of the
documentation were also updated and some extra information given.

Another important thing is that we wanted this line width to be the same
over the programming, documentation, and commit message linewidths, so it
doesn't cause confusion. A line-width larger than 75 in commit messages
will be too large when running `git log' (with full commit message).

Finally, note that all line widths should (can) not be immediately changed
to the new value, since it would completely ruine the version control
history. Instead, we hope that gradually (with more portions of the source
being updated), the new line width will become common throughout all of
Gnuastro's source.

Until now, Gnulib would just be used to replace system-wide headers when
necessary, but with the inclusion of its nproc.c and nproc.h files, we are
now also using Gnulib specific libraries. So it was explained that such
headers should be put beside other imported libraries (like GSL) and all
such cases were corrected in the ui.c files.