dev: Categories in commit title, workflow updated
The full category name was too large in the commit titles (as seen from the
previus few commits). So they were changed to a shorter version to allow
more space for the title and also make reading the messages easier in `git
log --oneline'.

The workflow text was also updated (mainly to recall the copyright
assignment) and also to for clarity.