dev: Set indent-tabs-mode to t for makefiles in dir-locals.el
In makefiles it’s important to have `indent-tabs-mode' set to t in
Emacs.  Another strategy could be to selectively set `indent-tabs-mode'
to nil only in relevant modes (like `c-mode' and `texinfo-mode') and not
forcing `indent-tabs-mode' to nil globally, but for the time being there
are fewer modes that require setting `indent-tabs-mode' to t than the
contrary.

This fixes bug#47864.