No categories in commit title guidelines
The commit guidelines no longer require a "Category: " in the title. This
could cause confusion and was an extra burden and complication on the
developers when making commits while also taking up valuable space in the
short space available for a commit title. Thank you Mosè for your
well-written argument in the gnuastro-devel mailing list:

http://lists.gnu.org/archive/html/gnuastro-devel/2016-05/msg00019.html

It was also suggested that the body should start with the points relevant
to a user of Gnuastro, for example an option or argument update in a
utility or library (once libraries are also installed), since we will use
the logs to build the NEWS file on each release and also people who will be
pulling from the main repository can see the relevant points faster.