Updated Acknowledgements section and THANKS file
The Acknowledgements section in the book was updated to be more clear and
clearly separate the authors from the people list there. Also the role of
the institutions (and the fact that they disclaimed their copyright
ownership) was further highlighted.

The THANKS file was also similarly updated to have the full list of
indirect contributors (people who helped with ideas and suggestions that
have not directly gone into the version controlled history).