Incorrect double-quotations in bootstrap.conf removed
In the commit which introduced the curl program when wget is not available
(commit hash 5610289), the directory for the Autoconf archives was also set
using ../ instead of a SED operator. However, the full arguments to `cp'
were also placed within quotation signs, so `cp' would complain about no
destination file and abort. The quotation marks were removed here.