gendocs.sh now comes from Gnulib
The gendocs.sh script (which is used by the maintainer to update the
Gnuastro documentation webpage) was previously copied into the repository
from Gnulib and tracked (version controlled).

From now on, it will be imported from Gnulib like all the other
bootstrapped files that were not written as part of Gnuastro.