Make authors.texi a double-colon rule
authors.texi rule has no prerequisites, so once the file is created (and
it’s created empty when running make in top directory) the rule is never
run again afterwards, because the target appears always updated.  The
double-colon rule is always executed when there are no prerequisites.
Note that because of this, also every rules depending on authors.texi
will always be executed, but this is a necessary side effect since
authors.texi should be updated after any change to git history (and
cannot depend on "../.git" because this isn’t available in non-git
snapshots).  However, better always than never.

In addition, don’t count merge commits.