Curly braces added for code clarity and problem
Thanks to the updated GCC 6.1's warnings, I found out that in one case (the
change in src/noisechisel/label.c) I had missed braces! This was a strange
situation, because the indentation was correct, but the braces were
missing! This was corrected.

The other GCC 6.1 warning in src/convertt/eps.c was not a problem in the
coding, but some curly braces could help clarify the situation. The new
line is printed in the same line to ease reading (and later modifying or
debugging) the EPS source.