Fix genauthors scripts
As already explained in message of commit
  * af961a7 (2016-05-08)  Make authors.texi a double-colon rule
authors.texi is created empty when running "make" in top directory (in
particular "make -jN", with N > 1), so this file needs to be generated
even if it’s more recent than .git directory.  There is little overhead
in recreating the file, it’s better doing this twice than none.

In addition, repeat again the same fix to sed command made with commit
  * 58e1cf4 (2016-05-07)  Fixes to documentation

Don’t count merge commits also in genauhtors script in top directory.