doc/genauthors now works with different build directory
The doc/genauthors script could only be built when the source and build
directories where identical. So `make distcheck' would complain about not
finding the script. When the script was found, it couldn't find the
existing authors.texi file (because it was in the source directory). So now
it takes the source and build directories as input and uses them in its
laws.

Also a default result was added (to be rarely updated, since it is not a
common situation), so if the authors.texi could not be found, the authors
until a specific version could be used.