doc/authors.texi only built in configure time
Creation of the Texinfo book source depends on doc/authors.texi. Until now
doc/authors.texi was a double colon target with no prerequisits, meaning
that it would be built when ever the manual needed to be built. So in `make
distcheck', where we don't have write permissions in the source directory
(which contains doc/authors.texi), this would cause problems.

A similar problem was reported before in the link below:

https://lists.gnu.org/archive/html/automake/2015-03/msg00070.html

The first solution proposed to this problem is possible, but is a little
dirty/complex (involves temporary files). However the final answer is more
elegant. In practice, we only want this file (and thus the manual authors)
updated when a new version is released (official or unofficial), when the
developers are working for themselves, there is no need to update the
authors list.

Fortunately we have recently adopted the automatic version numbering (from
Gnulib's git-version-gen). In the book we have also advised everyone to
generate a new version if they plan to use their cloned/developed version
in any data analysis, or paper, or if they want to share their work (by
running `autoreconf -f'). That procedure will update `./configure' and thus
regenerate the authors list for the manual too.

This fixes bug #47878.