AUTHORS is now double colon rule, not .PHONY
Until now the AUTHORS file was a .PHONY target (to ensure that it is
re-built everytime independent of it being already present or not). But the
double colon rule has a similar outcome and is more elegant (takes one
extra line from the Makefile.am, which could also cause bugs).