Configure message for Ghostscript version check
A configure message was added to the Ghostscript version check let the user
know what Ghostscript is doing (checking the version of Ghostscript) and
the final version number. This was done with Autoconf's AC_MSG_CHECKING and
AC_MSG_RESULT before and after the version checking. Also the whole check
was placed in an AS_IF so it is only run when the actual Ghostscript
executable is found.

This can also be useful later, if we allow for different executable names
for Ghostscript.