Updated forwebpage script and explanations
The `.sh' suffix was removed from the forwebpage script and also the
instructions in doc/README were updated to reflect the new file
structure. It is mainly of interest for the maintainer. Previously all the
webpage files were stored in this directory. But since commit 23667f0
("Webpage: corrected forwebpage.sh and used external local CVS repo"), this
is no longer the case. I had forgot to update the doc/README file.