Added -u option to copy webpage contents
After building and correcting the webpage contents in `doc/forwebpage', the
contents would be copied to the destination directory. However, even if a
file was not changed the newer version would be copied, so CVS would upload
everything! The -u (--update) flag was added to the `cp' command in
doc/forwebpage so CVS only uploads the files that need changing.