Checking for webpage updates with diff
The previous commit (c83625f: Added -u option to copy webpage contents) was
not a solution to the problem state there, since all the generated manual
files were just created and had a newer time stamp. So the `doc/forwebpage'
script was updated to check for actual differences with the `diff' utility
and only copy files that actually have different contents.