Check version and date of online manual
The version number and date printed on the manual come from two different
sources: the version number comes from ./configure, while the date printed
on the manual is the modificationd date of doc/gnuastro.texi. This might be
useful for traditional version numberings (where the version number is
fixed, but the manual modification date is important). However, here that
we generate the version number automatically it could cause confusion: the
version number and date can not be consistent and thus confuse someone who
is trying to specify exactly which version's manual they are reading.

So a check was added to doc/forwebpage to see if the date of these two
files to avoid such issues. Also when doing the diff of the separate files,
the name of the file is written along with the (possible) diff result and a
statement saying what has happened to this file (was it copied or not).