Webpages with only version and date updates not uploaded
gendocs.sh adds a version number in the comments of the individual pages it
creates. We have also added a MONTH YEAR string to the end of the pages,
these two values can regularly change while the actual contents of the page
doesn't. So to avoid updating the pages that have not changed, these two
particular lines were ignored by `diff' in deciding which files it should
copy to the upload directory.

To ignore the date line, I added a HTML comment to the line which is unique
to this line. This allows us to later change the date format if we
want. The first one is already part of a comment, so we have no role in
creating it and it is a unique commented line.