GNU Autoconf archive version checker now used
The old check we had used for the version of Ghostscript would not work
when the version number was 9.05. Now we use the Autoconf archive macro
ax_compare_version.m4 to check for versions. This is a powerful version
checker that can also come in handy later if we need more such checks.

bootstrap.conf was also updated to copy/download this file.