VCS section moved to the Installation chapter
The "Version controlled source" section of the book was moved to the
"Installation" chapter from the "Developing" chapter. While developers will
certainly want version control, with the adoption of the automatic version
generation, users might also want to stay up to date. Having this section
in the "Developing" chapter would discourage such users. In the process the
full "Installation" chapter was also edited so this new section blends in
nicely.

During the work above some other issues were also addressed:

  - The "Requirements" sections are now changed to a single "Dependencies"
    section, which contains a "Mandatory dependencies", "Optional
    dependencies" and a new "Bootstrapping dependencies" subsection.

  - The "Install Gnuastro" section was also changed to "Build and Install"
    to be more clear and in coherent with the other sections of the
    "Installation" chapter.

  - The usage of the Texinfo VERSION variable was significantly decreased
    throughout the body of the book. It was replaced with `X.X' since it
    can be very annoying when the manual isn't an official release and
    people will probably be using unofficial releases too.

  - In the installation of CFITSIO a line was added to remove the test
    executables it builds before installation, we only need `fpack' and
    `funpack'.

  - The `TESTdir' name was changed back to the actual name of the testing
    directory (`tests/').

  - Several occurances of `your self' were corrected to `yourself'.