Notices added after make, make check and make install
Until now, a notice would only be printed after configure. Now there is a
similar message after each step, explaining what was done and what the user
should do next (so the message after configure just says to run
`make'). Since all the work is done in the sub-directories, these messages
will be printed in the end even if make is called to run multiple jobs
simultaneously.

As Gnuastro gets larger, the building process will take longer and these
messages can help the users along the way (especially the ones that are new
to the GNU Build System).