Release tarball subsection added in book
Previously only a URL was given for the release tarball, which was not
enough, since we will soon have prerelease (alpha) versions too. So A new
subsection was added to the "Downloading the source" section titled
"Release tarball", explaining the tarball download links and also the
mirrors users can use.