Temporary build webpage in another directory
The `doc/forwebpage' script is used by the maintainer to prepare and upload
the webpage updates. It can also be used by others to generate the webpages
on their local systems. Until now, the temporary directory where all the
contents were kept before moving to their final position was hardcoded in
the script and was in the doc/ directory, but now its a variable defined on
the start of the script.

Also now, the names of the files are only written when the contents
actually differ.