gnuastro-commits implemented in the book and webpage
The gnuastro-commits mailing list is a new feature that was kindly
suggested and implemented in the repository by Bob Proulx (a Savannah
maintainer). All the commits made to the official repository will now be
automatically sent to this mailing list for circulation to the Gnuastro
developer community. This new feature required some changes in the
documentation that have been implemented in this commit. This should be the
first commit that is forwarded to that mailing list.

One important addition for future commits is this: when a commit is
directly related to a bug or task, please indicate the bug or task ID in
the commit, even if your work is not yet finished. This will help
interested subscribers to the gnuastro-commits mailing list easily join the
discussion by finding the proper bug or task and posting a comment to it.

In the webpage, the downloading options were more clearly elaborated and
the ability to clone Gnuastro was also mentioned there.