Dependencies defined for tests
As described in bug #47957, the tests depend on each other. However, until
now, there was no specific definition of this dependence. Therefore `make
check' could not be run on multiple threads. Now dependencies are clearly
defined in `tests/makefile.am' as explained in the "Parallel test harness"
section of the Automake manual (using the `.log' outputs).