Tests will SKIP when input is not ready
In the previous commit (e3edcc9: Dependencies defined for tests), we
defined the dependencies between the tests, so they are run in the proper
order. However, the `.log' files will be created even if the tests
fails. So if one of the dependencies fails to make an image, the rest would
still fail (because there was no image). So now, all the tests have a
special content structure where their input is set in the top and the
executable and input image is checked before the actual script is run. In
this way, if a dependency of a test fails, it will be SKIPped, it won't
FAIL.

This fixes bug #47957.