Tags are now prefixed with v
Previously the tags were just the major and minor version number (for
example 0.0). Savannah's cgit web interface will create a tarball for each
tag point and keep it on the top page. So for example the tarball of the
0.0 tag would be: gnuastro-0.0.tar.gz. Note that this tarball is not
bootstrapped, it only contains the version controlled files.

This will cause confusion later. Because the bootstrapped tarball that is
ready for configuration will also have exactly the same name! Since the
cgit interface uses the exact tag string, this problem can be fixed if a
`v' is prefixed to the tag. So the tarball that is downloaded from there
will be gnuastro-v0.0.tar.gz. This will fix the problem. Also, it appears
to be very common to use the `v' prefix for tags, such that Gnulib's
git-version-gen script assumes a `v' prefix by default. So the `--prefix
""' option that was given to it in `configure.ac' is no longer necessary.