Cleaned the test dependencies
Until now, the tests that depended on a previous test also depended on
`prepconf.sh' (the script that will bring a copy of all the configuration
files for the tests). However, this was redundant, because this script is
already a prerequisite of the prerequisite tests. So it was removed to make
things more clear and less bugy. It is now only a prerequisite for tests
that don't depend on the other utility tests.