Tags prefixed with gnuastro_v
In a previous commit (122597e: Tags are now prefixed with v), the common
`v' prefix for tags. However, cgit was able to recognize this and
internally removed the `v' to make the tar.gz. So after some tests, a new
format was chosen: `gnuastro_vX.X', where the two Xs are the major and
minor version number. As explained in the new tag, with this format, cgit
will make a reasonably informative, somehow similar, but clearly different
tarball (`gnuastro_v0.0.tar.gz' instead of `gnuastro-0.0.tar.gz' which is
used in the FTP server).