Corrected imgcrop/imgpolygon.sh test script
This script would make an output file named `mkprofcat1_crop.fits', then
rename it to the desired output. This could cause troubles when run in
parallel because the same could be created by other scripts. So in this
test, now ImageCrop is instructed to directly write its output into
`imgpolygon.fits'.