Book update on institutions in copyright assignment
Previously the "Acknowledgments" and "Copyright assignment" sections of the
book assumed that every contributor would need a disclaimer from their
university, or employer. However, a careful reading of the maintainers
guide and `/gd/gnuorg/conditions.text' clearly that only when you are
employed to do programming, or have made such an agreement is it necessary.

So the list of institutions were deleted from the "Acknowledgments" section
and the explanations in "Copyright assignment" were updated along with a
cartouche explaining this special condition.