Ver. number generation removed from commit guidelines
Until now, the "Commit guidelines" section of the book contained two
sub-sections: "Commit message" and "Generate unique version number". This
was not the best place for the latter, it is more closely related to the
"Synchronization" section of the book. So it was removed.

With this subsection removed, the separate section title "Commit message"
title was also removed and its text is now part of the general "Commit
guidelines" section.