doc/fdl.texi is no longer under version control
The GNU Free Documentation License (FDL) is included within Gnulib,
therefore it was wrong to keep it under version control in Gnuastro. With
this commit, it is no longer tracked and is imported from Gnulib.

There is a problem though. As far as I have understood so far, including
the following command in the `doc/Makefile.am' should have been enough to
easily build Gnuastro with the `fdl.texi' file in the `bootstrapped/doc'
directory.

    AM_MAKEINFOFLAGS = -I $(top_srcdir)/bootstrapped/doc

While this does work for Info and HTML outputs, for PDF and DVI outputs it
fails. So for the time being, the bootstrap process puts a copy of
`fdl.texi' within the `doc/' directory to do the job. If a more elegant
solution can be found such that this copy is not needed and all
documentation can be built with `fdl.texi' in the `bootstrap/doc/'
directory, it would be much better. A mail was sent to bug-automake to see
if this is a bug worth fixing or not.