fdl.texi is now kept in the bootstrapped directory
In an effort to keep the `doc/' directory as clean (from non-hand-written
files) as possible, a work-around is used to keep the `fdl.texi' file in
the `bootstrapped/doc' directory instead of having to copy it in `doc/'
during the bootstrapping process (as done in e80944a: doc/fdl.texi is no
longer under version control).

This process was inspired by the solution to Automake bug report #23599
which I submitted, Mathieu Lirzin suggested overriding the value of
`TEXI2DVI' which is a nice solution. However for some reason, Automake
complains about the value substitution and exits with an error. So what I
have done now is to manually (using SED) correct the value of TEXI2DVI at
the end of the bootstrap process in bootstrap.conf's `bootstrap_epologue'.

This is not a good solution, because if automake is ran outside of the
bootstrap process (for example when editing the `Makefile.am's), then it
will not be able to build PDF and DVI outputs.