doc/genauthors works on a separate build directory
Until now, the `doc/genauthors' script was only set up to work when the
build and source directories where the same. If they were different, this
script would fail to find a Git repository. So it now takes in an argument
(the `top_srcdir' variable) and uses it in checking for a `.git' directory
and importing the `.mailmap' file. It then uses them with Git to build the
authors.texi file in the output directory.