Better workaround for fdl.texi
The previous work-around for `fdl.texi' could cause big problems when
Automake was ran outside of the `bootstrap' script. For a more reliable
workaround, we are now using the `AM_V_texinfo' variable. This variable
appears to have been defined for verbosity, but since it isn't a
significant issue (at least compared to this), we have replaced it with
this necessary flag.

Later (when Autoconf's bug #23599) is fixed, we can remove this definition
entirely, since `AM_MAKEINFOFLAGS' should be enough. But until then, we
will need to use this working alternative.