No more warning errors from Automake
Thanks to Mathieu Lirzin from Automake, the override problem mentioned in
commit 0394025 (fdl.texi is now kept in the bootstrapped directory) was
found to be due to the `-Werror' option that was passed to Automake in
`configure.ac'. So the workaround mentioned in commit 5d8967c (Better
workaround for fdl.texi) is no longer necessary.

To avoid such issues raising again, the -Werror option was removed from
`AM_INIT_AUTOMAKE' in `configure.ac'. All warnings will be printed and will
thus be clearly visible, however, there will be no error.