New "Branching workflow tutorial" subsection in book
This new section explains all the necessary steps for someone to easily be
able to apply their changes to ultimately include into Gnuastro. During the
process, the "After making changes" section was also moved to the end of
the "Developing" chapter and some of the text in "Contributing to Gnuastro"
was edited.