Removed "Table manipulation" section in book
From the very early days, this chapter was supposed to be a place holder to
add table manipulation utilities to Gnuastro too. Later on, the Savannah
project management page was set up and we have defined task #13579 for the
same purpose. Until work on that task starts, the few lines that were
devoted to this empty chapter have been removed from the manual.

A minor correction was also made in the hash explanation in "Version
numbering". Some left-over corrections from the previous commit (eb19945:
Spell check done on the book) were also applied.