No log files generated during tests
When the tests are run on multiple threads (with `make check -jN', N>1),
the log files generated by some of the utilities can cause complications
and thus failures. The cause of the problem is that currently the log files
of each utility have the same name, irrespective of their input.

Therefore to avoid such wrong FAIL results, all log files are disabled in
the tests using the configuration files that are set in the
`tests/prepconf.sh' script (first test).