Updates to the installation sections of the book
The "Quick start" section and the "Installation" chapter were reviewed and
edited to be more clear and link more closely with each other. Since these
pages will probably show up in internet search results, it is important
that they be closely connected to users can easily navigate trough the
installation.