Updated explanations in THANKS
Following commit 608cf48 (Book update on institutions in copyright
assignment), the explanations in the THANKS file were also updated. Here,
the university names were kept to acknowledge their constructive role in
the developers work, but the discussion about copyright was removed. We'll
test this here and if it is later agreed, we can follow the same style in
the book: acknowledging the role of our host institutions, mainly for the
environment and support they provided to us (if indeed we feel they did!).