Updated list of utilities in the book
The explanation of the utilities in the book was originally very
short. Originally, it was intentional, so the full explanation would fit
into one line. But that would be too terse. So the explanations from the
'README' file were taken and put here. During the process a small
correction to the 'README' file was also made.