Added explanation on the periodic nature of e^{ix}
In the "Circles and the complex plane" section, we had just stated that
e^{ix} is periodic. Two paragraphs were added with a link to Wikipedia to
show it geometrically, this is a very important (and hard to visualize)
concept that is important in correctly understanding and thus using the
Fourier space, so it was added here.