ChangeLog now explains how it is no longer updated
The ChangeLog update was stopped in each commit following the discussion in
task #13779. However, this choice was not reflected in the ChangeLog its
self. So if someone would open the ChangeLog, they would think the last
change was on November 5th, 2015! To avoid such confusions, an item was
added to the top of the ChangeLog explaining how we have stopped updating
it. Two new tasks have thus been defined:

 - task #14007: to automatically generate the ChangeLog from the version
   controlled history.

 - task #14008: to discuss if it is good to adopt the ChangeLog formatting
   in the commit messages.

This finishes task #13779.