Webpage checkout and temporary dirs on command-line
The `doc/forwebpage' script is in charge of making and preparing the
webpage files (in particular the book webpages). Until now the temporary
and final directories were hard-wired into the script, but now they are
read as two argumens to the script. Also if the temporary directory already
exits, it will not attempt to completely delete it automatically, it will
warn the user and abort.