New "Forking tutorial" name for section, more clarity
The old "Branching workflow tutorial" was renamed to "Forking tutorial" to
be more clear. Also following some questions asked by Navid Mousavi, some
further clarifications were added to the "Developing" chapter:

  - That interested people don't need to become a Savannah member to be
    able to clone and develop in Gnuastro.

  - That the repository hosting services will need to know your public SSH
    key in order for the `git push' command in the "Forking tutorial"
    section to work.

Besides these, in some places, it was found that "let's" was mistakenly
written as "lets". This was also corrected.