Added links and explanation in "Report a bug"
Links to two nice essays on best ways to report a bug report were included
in the "Report a bug" section. I found these links in the Automake manual
and found them to be very useful. "Be descriptive" was also added as a
suggestion for a good bug report and some minor typos were corrected.