Library header files moved into lib/gnuastro
The library header files were moved into the new directory `lib/gnuastro'
from the old `include/' directory. This was done in preparation to have the
libraries become installable (task #13765). The basic idea is that the
standard place for a large group of header files to be installed is the
'$(pkgincludedir)' of Automake. This is a $(PACKAGE) (or `gnuastro/')
directory within the $(includedir).

We want the code inside of Gnuastro to be as similar as possible to the
code someone would wright using Gnuastro for their own purposes in their
own programs. On the other hand, it would be confusing to change the old
`include/' directory to `gnuastro/', since it was in the top Gnuastro
directory. So the new `gnuastro/' directory was moved within the `lib'
directory and a `README' file was also included to explain the purpose of
this (vaguely named when viewed within Gnuastro's source) directory. On the
other hand after installation, it will be very clear in the installed
`include/' directory.

As discussed in the coding conventions, the `gal_' prefix was not added to
the header file names. To re-iterate: the headers depend on each other
(like GSL). So in the header files you will also see `#include
"gnuastro/headername.h"'. If the user decides to install the header files
in the same directory as other programs (having no `gnuastro' directory),
Gnuastro will fail to compile (GSL's header files also depend on each other
in a similar way). Therefore this assumption is hard-wired into the source
code. It might be possible to use `config.h' macros set at configure time
to do this, but that is not common and will make the code harder to read.