Unnecessary include flag (-I) removed in all Makefile.ams
The `-I$(top_builddir)/bootstrapped/lib' flag that was given to
`AM_CPPFLAGS' in all the Makefiles has been removed. Gnulib's header files
are all in the `$(top_srcdir)/bootstrapped/lib' directory, not in the
`builddir'. Including this was a mistake some time ago, and since it didn't
cause any errors, it was never removed. It might cause confusion (leading
an unexpecting reader to think that some of the header files are generated
by the build process, while they aren't).

Also, the spacing from the top comments of the `Makefile.am's was not
uniform, so the top uncommented line of the `Makefile.am's was set to be 5
lines below the comment.