configure is now a prerequisite of .version
Following `git-version-gen''s suggestion, until now `.version' did not
depend on anything. So the only way to update it was to delete it and run
`make'. This manual operation could cause confusion if forgotten (for
example in unofficial releases). So now `.version' depends on the
`configure' script. Any time `configure' is rebuilt (and thus the version
is updated), the `.version' file is also updated.