doc/authors.texi defined as BUILT_SOURCES
When Make was run on multiple threads, the `doc/authors.texi' file would
sometimes be empty (bug #47987). So it was defined as a `BUILT_SOURCES' in
Automake. All targets specified as `BUILT_SOURCES' are built before
anything else (so the parallel building will be done after them).

On my GNU/Linux system, this seems to have solved the problem and
`doc/authors.texi' is built correctly in multithreaded builds that failed
before. So I will mark the bug as fixed ("Closed" on Savannah), however, if
this solution is not universal, we can change the status of this bug back
to "Open".

This fixes bug #47987.