AUTHORS now contains the version
Until now the `AUTHORS' file did not have the version of Gnuastro that it
referred to. So `genauthors' now reads the `.version' file and includes the
version in the produced AUTHORS file.

In `genauthors', tests were set to report unused and empty variable values,
also two typos were corrected in the generated AUTHORS file.