.version is prerequisite of AUTHORS
Following commit 2e0cfea (configure is now a prerequisite of .version), the
`.version' file is now a prerequisite of `AUTHORS' (`.version' is used by
the `genauthors' script to add the version number to `AUTHORS'). Since
`.version' also depends on `configure', so AUTHORS will now be generated
every time that Autoconf re-builds the `configure' script. During
development, updating the `configure' scripot is not too common, so this
won't slow down the process.