Man pages can be made in parallel
Until now, the man pages did not depend on the actual executable of each
utility. Therefore, when run in parallel, they would be build along side
the utility and all its sources, and therefore would fail (since `help2man'
needs the executable). With this commit, the executable is also a
dependency of the Man page, so there won't be any problem in parallel
builds.

`help2man' is run by calling the executable string along with the two
`--output' and `--source' options. Their order does not matter. So in this
update, they were ordered such that the `--source' option (which can be
long) is in a new line. This makes it much more likely for both lines to
fit nicely in an 80-character wide terminal emulator.

Since the man page is now dependent on the executable, it is also be built
in the build directory (when `help2man' is present), but it is not cleaned,
so it was also added to Automake's `CLEANFILE' variable (without this,
`make distcheck' would fail).