tests/.gnuastro directory is now cleaned
The `tests/.gnuastro' directory is built by the `tests/prepconf.sh' script
to keep all the configuration files for testsing. However, to clean it, we
had relied on Automake's `CLEANFILES' variable. As the name suggests, this
variable is only built for files, not directories. So the build directory
would not get cleaned. Spotting this problem was hard because Automake
doesn't abort in such cases and the notice was too deeply buried in the
other outputs.

This problem was fixed using Automake's `clean-local' rule.