gendocs in bootstrapping, no emails parts lists
The call to `genauthors' in `bootstrap.conf' didn't have the recently
added second argument. This was added to allow the generation of AUTHORS
initially during bootstrapping. At the bootstrapping stage there is also no
`.version' file, so `genauthors' will use `git describe'.

Also the emails were removed from the list of authors of the separate parts
of Gnuastro to be more clear. They are already listed (alphabetically) in
the aggregate list.